+++ src/config.c
@@ -580,6 +580,9 @@ void config_save(int is_global, char *fn)
 	section_t *current_section;
 	list_t *head = is_global ? &global_config_head : &machine_config_head;
 
+	if (!f)
+		return;
+
 	current_section = (section_t *)head->next;
 
 	while (current_section)